requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-dialogs>=4.14.0
botbuilder-ai>=4.14.0
datatypes-date-time>=1.0.0.a2
BeautifulSoup4
validators
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-speech
pydub